Move ProtocolFeatures to scylla-cql-core
Note that it became public in scyllas lib.rs, as it should be.
